fixed some issues outlined by sonar
